clarify install steps and fly.io notes
